backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pandas
python-dotenv
pydantic